+++ pkg/reconciler/workload/basecontroller/controller.go
@@ -36,7 +36,7 @@ import (
 	kcpclient "github.com/kcp-dev/kcp/pkg/client/clientset/versioned"
 	apiresourceinformer "github.com/kcp-dev/kcp/pkg/client/informers/externalversions/apiresource/v1alpha1"
 	workloadinformer "github.com/kcp-dev/kcp/pkg/client/informers/externalversions/workload/v1alpha1"
-	"github.com/kcp-dev/kcp/pkg/reconciler/apiresource"
+	"github.com/kcp-dev/kcp/pkg/reconciler/apis/apiresource"
 )
 
 const GVRForLocationInLogicalClusterIndexName = "GVRForLocationInLogicalCluster"
